requirements.txt
Flask==3.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
itsdangerous==2.1.2
click==8.1.7
pandas
networkx
matplotlib
datasets
huggingface_hub
python-dotenv
